AICodeStudio/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "google-genai>=1.29.0",
    "gtts>=2.5.4",
    "langdetect>=1.0.9",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.1",
    "sift-stack-py>=0.8.2",
]
